+++ src/panel_window_dui.cpp
@@ -145,11 +145,11 @@ public:
 	{
 		if (p_what == ui_element_notify_colors_changed)
 		{
-			PostMessage(m_hwnd, callback_id::on_colours_changed, 0, 0);
+			PostMessage(m_hwnd, static_cast<unsigned int>(callback_id::on_colours_changed), 0, 0);
 		}
 		else if (p_what == ui_element_notify_font_changed)
 		{
-			PostMessage(m_hwnd, callback_id::on_font_changed, 0, 0);
+			PostMessage(m_hwnd, static_cast<unsigned int>(callback_id::on_font_changed), 0, 0);
 		}
 	}
